Login page now matched SignUp page from a style perspective:
