+++ frontend/src/pages/Login/LoginPage.css
@@ -1,94 +1,13 @@
 .home {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: flex-start;
-    height: 100vh;
-    width: 100vw;
-    padding: 20px;
-    font-family: 'Arial', sans-serif;
-    background-color: rgb(214, 243, 216);
-  }
-  
-  .header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 97%;
-    padding: 15px 40px;
-    background-color: #333;
-    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.15);
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    z-index: 1000;
-    border-radius: 0 0 12px 12px;
-  }
-  
-  .logo-container {
-    display: flex;
-    align-items: center;
-  }
-  
-  .logo {
-    height: 55px;
-    margin-right: 15px;
-  }
-  
-  .title {
-    font-size: 28px;
-    font-weight: bold;
-    color: #f4f4f4;
-    text-transform: uppercase;
-    letter-spacing: 1px;
-  }
-  
-  .nav {
-    display: flex;
-    gap: 20px;
-  }
-  
-  .nav-button {
-    border: none;
-    padding: 10px 20px;
-    border-radius: 6px;
-    background-color: #19a71c;
-    color: white;
-    text-decoration: none;
-    font-size: 16px;
-    font-weight: bold;
-    transition: background-color 0.3s, color 0.3s;
-  }
-  
-  .nav-button:hover {
-    background-color: #f0e919;
-    color: #333;
-  }
-  
-  .main-content {
-    text-align: center;
-    margin-top: 140px;
-    width: 70%;
-  }
-  
-  .welcome-box {
-    border: none;
-    border-radius: 10px;
-    padding: 25px;
-    background-color: white;
-    box-shadow: 0 3px 10px rgba(0, 0, 0, 0.1);
-    max-width: 800px;
-    font-size: 20px;
-    font-weight: bold;
-    color: #333;
-    margin: 20px auto;
-  }
-  
-  .login-title {
-    font-size: 48px;
-    font-weight: bold;
-    color: #333;
-    margin-bottom: 20px;
-    padding-bottom: 30px;
-  }
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
+  height: 100vh;
+  width: 100vw;
+  padding: 50px;
+  padding-top: 180px;
+  font-family: 'Arial', sans-serif;
+  background-color: rgb(214, 243, 216);
+}
+
