+++ frontend/src/pages/Login/LoginPage.jsx
@@ -1,59 +1,13 @@
-import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import { login } from "../../services/authentication";
 import Header from "../../components/Post/Header";
+import Login from "../../components/Post/Login";
 
 export const LoginPage = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const navigate = useNavigate();
-
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-    try {
-      const token = await login(email, password);
-      localStorage.setItem("token", token);
-      navigate("/posts");
-    } catch (err) {
-      console.error(err);
-      navigate("/login");
-    }
-  };
-
-  const handleEmailChange = (event) => {
-    setEmail(event.target.value);
-  };
-
-  const handlePasswordChange = (event) => {
-    setPassword(event.target.value);
-  };
-
-return (
+  return (
     <div className="home">
       <Header />
-      <main className="main-content">
-      <h1 className="login-title">Login</h1>
-      <form onSubmit={handleSubmit}>
-        <label htmlFor="email">Email:</label>
-        <input
-          placeholder="Email"
-          id="email"
-          type="text"
-          value={email}
-          onChange={handleEmailChange}
-        />
-        <label htmlFor="password">Password:</label>
-      <input
-        placeholder="Password"
-          id="password"
-          type="password"
-          value={password}
-          onChange={handlePasswordChange}
-        />
-        <input role="submit-button" id="submit" type="submit" value="Submit" />
-      </form>
-      <p className="signup-text"> Don&apos;t have an account? <a className="loginButton" href="/signup">Sign Up Here</a></p>
-      </main>
+      <Login />
     </div>
   );
 };
+
+export default LoginPage;
